Back button - DONE
I have managed to add a back button. They way I did it was to create another component and user is redirected to it when clicks on the search button. I used routes.push to change the url.
After that the error message did not work and I added another conditional to check if there is data in the movies array and show error if no data.
Also checked my code and made sure I add comments and indent correctly