Added pagination
I added pagination to the results page
I allow up to 5 pages to show based on the total number of results - if more than 40 - I show 5 pages which is the maximum as I don't think more than that is relevant. If less results it should show the number of pages accordingly
I achieved that by adding variable pageNumber and totalResults
I get the totalResults from the API and the use it in the template to make the pages dynamic
I added pageNumber as variable and set it to 1 as a start. After that I use it as a parameter to create another function which will call the API with the corresponding page required and return the data from this page. The API divides the results in batches by 10